fix issue with mathjax support
This partially addresses #60.

We were doing it incorrectly all along. The mathjax script needed to be
in the *footer*, not the header. I know this is not exactly the right
way to do it since I'm copying over the footer from the layouts instead
of specifying a foot_custom.html, but I'll work on it later.